Split protocol out
