+++ src/coffea/_compute.py
@@ -6,14 +6,10 @@ from threading import Condition, Thread
 import time
 from typing import Callable, Iterator, Literal, Protocol, Sized, TypeVar
 
+from coffea.compute.protocol import Computable, ResultType
+
 
 T = TypeVar("T")
-
-
-class ResultType(Protocol):
-    def __add__(self: T, other: T, /) -> T:
-        """Merge two results together."""
-        ...
 
 
 class EventsArray(Sized, Protocol):
@@ -31,20 +27,6 @@ class ProcessorABC(ABC):
     def process(self, events: EventsArray) -> ResultType:
         """Process a chunk of events and return a result."""
         raise NotImplementedError
-
-
-class Computable(Protocol):
-    def __iter__(self) -> Iterator[Callable[[], ResultType]]:
-        """Return an iterator over callables that each compute a part of the result.
-
-        This establishes a global index over the computation. We can use filters
-        and slicing on this basis to create resumable and retryable computations.
-
-        A potential optimization is to yield a tuple of (index, callable) so that
-        the index can be smarter than just the integer position in the iterator.
-        Then Computable could be a Container type.
-        """
-        ...
 
 
 @dataclass(slots=True)
@@ -112,35 +94,6 @@ class Dataset(StepIterable):
         if self.traversal == "breadth":
             raise NotImplementedError
         return chain.from_iterable(map(File.iter_steps, self.files))
-
-
-class Task(Protocol):
-    def result(self) -> ResultType:
-        """Get the full final result of the computation."""
-        ...
-
-    def partial_result(self) -> tuple[ResultType, Computable]:
-        """Get a partial result and the corresponding continuation computation.
-
-        The partial result may either be because the task is not yet complete,
-        or because the computation failed on a subset of the data.
-
-        TODO: add cancel parameter to indicate whether to stop ongoing computation?
-        """
-        ...
-
-    def wait(self) -> None:
-        """Block until the computation is complete.
-
-        TODO: add timeout parameter and return_when options?
-        c.f. https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.wait
-        FIRST_COMPLETED, FIRST_EXCEPTION, ALL_COMPLETED
-        """
-        ...
-
-    def done(self) -> bool: ...
-
-    def cancel(self) -> None: ...
 
 
 @dataclass
@@ -303,19 +256,6 @@ class SimpleTask:
                     self._output = self._output + result
                     self._index += 1
         return self._set_done()
-
-
-class Backend(Protocol):
-    def compute(self, item: Computable) -> Task:
-        """Launch a computation and return a Task representing it.
-
-        The backend holds any resources needed to perform the computation,
-        such as a thread pool, process pool, or cluster connection. It should
-        manage a queue of tasks and execute them in FIFO order.
-
-        Probably this should have a context manager interface to clean up resources.
-        """
-        ...
 
 
 def _work(task_queue: Queue[SimpleTask]) -> None:
